+++ pkg/object/object.go
@@ -33,11 +33,11 @@ type Object interface {
 
 // TODO: rename this to num to match evy spec
 type Integer struct {
-	Value int64
+	Value float64
 }
 
 func (i *Integer) Type() ObjectType { return INTEGER_OBJ }
-func (i *Integer) Inspect() string  { return fmt.Sprintf("%d", i.Value) }
+func (i *Integer) Inspect() string  { return fmt.Sprintf("%f", i.Value) }
 func (i *Integer) HashKey() HashKey {
 	return HashKey{Type: i.Type(), Value: uint64(i.Value)}
 }
